Add files via upload
Tic-tac-toe is a two player game (one of them being the computer program). The two players take turns putting marks on a 3x3 board. The player who first gets 3 of his/her marks in a row (vertically, horizontally, or diagonally) wins the game, and the other loses the game. 
This project helped me to experience real life projection of the knowledge and how linear data structure 
such as “Arrays” and Back tracking Algorithms.